+++ app/controllers/transfers/projects_controller.rb
@@ -57,7 +57,11 @@ class Transfers::ProjectsController < ApplicationController
       :incoming_trust_sharepoint_link,
       :outgoing_trust_sharepoint_link,
       :assigned_to_regional_caseworker_team,
-      :handover_note_body
+      :handover_note_body,
+      :two_requires_improvement,
+      :inadequate_ofsted,
+      :financial_safeguarding_governance_issues,
+      :outgoing_trust_to_close
     )
   end
 end