+++ spec/forms/transfer/create_project_form_spec.rb
@@ -101,6 +101,30 @@ RSpec.describe Transfer::CreateProjectForm, type: :model do
         expect(form.errors.messages[:incoming_trust_ukprn]).to include I18n.t("errors.attributes.incoming_trust_ukprn.ukprns_must_not_match")
       end
     end
+
+    describe "inadequate Ofsted" do
+      it "cannot be blank" do
+        form = build(:create_transfer_project_form, inadequate_ofsted: "")
+
+        expect(form).to be_invalid
+      end
+    end
+
+    describe "financial, safeguarding or governance issues" do
+      it "cannot be blank" do
+        form = build(:create_transfer_project_form, financial_safeguarding_governance_issues: "")
+
+        expect(form).to be_invalid
+      end
+    end
+
+    describe "Outgoing trust to close" do
+      it "cannot be blank" do
+        form = build(:create_transfer_project_form, outgoing_trust_to_close: "")
+
+        expect(form).to be_invalid
+      end
+    end
   end
 
   describe "urn" do
@@ -184,6 +208,14 @@ RSpec.describe Transfer::CreateProjectForm, type: :model do
     end
   end
 
+  describe "two requires improvement" do
+    it "cannot be blank" do
+      form = build(:create_transfer_project_form, two_requires_improvement: "")
+
+      expect(form).to be_invalid
+    end
+  end
+
   describe "region" do
     it "sets the region code from the establishment" do
       project = build(:create_transfer_project_form).save
@@ -252,6 +284,20 @@ RSpec.describe Transfer::CreateProjectForm, type: :model do
         expect(Note.count).to eq(1)
         expect(Note.last.body).to eq("This is the handover note.")
         expect(Note.last.task_identifier).to eq("handover")
+      end
+
+      it "updates the tasks data with the other values" do
+        form = build(
+          :create_transfer_project_form, inadequate_ofsted: true,
+          financial_safeguarding_governance_issues: true,
+          outgoing_trust_to_close: true
+        )
+        form.save
+
+        expect(Transfer::TasksData.count).to be 1
+        expect(Transfer::TasksData.first.inadequate_ofsted).to be true
+        expect(Transfer::TasksData.first.financial_safeguarding_governance_issues).to be true
+        expect(Transfer::TasksData.first.outgoing_trust_to_close).to be true
       end
     end